requirements.txt
streamlit
sentence-transformers
groq
pandas
faiss-cpu
pymupdf
altair